+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 - Upcoming changes ...
+### Added
+- Added real-time URL validation with pathname removal and warning message
 
 ## [1.25.0] - 2025-11-05
 ### Added